import/migrate/amazon_aurora/requirements.txt
GQLAlchemy==1.7.0
mysql-connector-python==8.2.0 